// admin/src/generated-sources/openapi/api.ts
export interface BlogDto {
    /**
     * 
     * @type {string}
     * @memberof BlogDto
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof BlogDto
     */
    'uri'?: string;
    /**
     * 
     * @type {string}
     * @memberof BlogDto
     */
    'flag'?: string;
    /**
     * 
     * @type {number}
     * @memberof BlogDto
     */
    'user'?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof BlogDto
     */
    'tags'?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof BlogDto
     */
    'createTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof BlogDto
     */
    'updateTime'?: string;
    /**
     * 
     * @type {number}
     * @memberof BlogDto
     */
    'categorie'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDto
     */
    'published'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BlogDto
     */
    'commentable'?: boolean;
}